账单模块，对接好接口文档

+++ src/main/java/com/zhiji/smartbook/module/report/vo/ReportSummaryVO.java
@@ -1,17 +1,18 @@
 package com.zhiji.smartbook.module.report.vo;
 
 import lombok.Data;
+import java.math.BigDecimal;
 
 @Data
 public class ReportSummaryVO {
     private String periodType;
     private String startDate;
     private String endDate;
-    private Double expenseTotal;
-    private Double incomeTotal;
-    private Double avgDailyExpense;
-    private Double largestExpenseAmount;
+    private BigDecimal expenseTotal;
+    private BigDecimal incomeTotal;
+    private BigDecimal avgDailyExpense;
+    private BigDecimal largestExpenseAmount;
     private String largestExpenseCategoryName;
     private String largestExpenseMerchantName;
-    private Double expenseChangeRate;
+    private BigDecimal expenseChangeRate;
 }
